+++ packages/core-components/src/lib/atoms/index.js
@@ -5,4 +5,5 @@ export * from './button';
 export { default as CopyButton } from './copy-button/CopyButton.svelte';
 export * from './query-load';
 export * from './inputs';
+export * from './flowchart';
 export * from './grid';