+++ packages/core-components/src/lib/atoms/query-load/QueryLoad.svelte
@@ -25,9 +25,9 @@
 	<slot loaded={data} />
 {:else if !data?.__isQueryStore}
 	<!-- data prop was provided, but it is not a query store -->
-	{#if !data?.__isQueryStore && isEmptyDataset(data) && $$slots.empty}
+	{#if (Array.isArray(data) || !data) && isEmptyDataset(data) && $$slots.empty}
 		<!-- handle case where data is not a query store but is also empty -->
-		<slot name="empty" />
+		<slot name="empty" loaded={data} />
 	{:else}
 		<!-- Not a query store, nothing to be done -->
 		<slot loaded={data} />
@@ -41,10 +41,10 @@
 	</slot>
 {:else if _data.error && $$slots.error}
 	<!-- loading data returned an error -->
-	<slot name="error" />
+	<slot name="error" loaded={_data} />
 {:else if isEmptyDataset(_data) && !_data.error && $$slots.empty}
 	<!-- data loaded successfully, but the dataset is empty-->
-	<slot name="empty" />
+	<slot name="empty" loaded={_data} />
 {:else}
 	<!-- data loaded successfully -->
 	<slot loaded={_data} />