+++ packages/core-components/src/lib/atoms/shadcn/dialog/dialog-description.svelte
@@ -1,0 +1,13 @@
+<script>
+	import { Dialog as DialogPrimitive } from 'bits-ui';
+	import { cn } from '$lib/utils';
+
+	/** @type {string | undefined | null} */
+	let className = undefined;
+
+	export { className as class };
+</script>
+
+<DialogPrimitive.Description class={cn('text-sm text-gray-500', className)} {...$$restProps}>
+	<slot />
+</DialogPrimitive.Description>