+++ packages/core-components/src/lib/atoms/shadcn/range-calendar/range-calendar-cell.svelte
@@ -1,0 +1,23 @@
+<script>
+	import { RangeCalendar as RangeCalendarPrimitive } from 'bits-ui';
+	import { cn } from '$lib/utils';
+
+	/** @type {any} */
+	export let date = undefined;
+
+	/** @type {string | undefined | null} */
+	let className = undefined;
+
+	export { className as class };
+</script>
+
+<RangeCalendarPrimitive.Cell
+	{date}
+	class={cn(
+		'relative p-0 text-center text-sm focus-within:relative focus-within:z-20 [&:has([data-selected])]:bg-gray-100 first:[&:has([data-selected])]:rounded-l-md last:[&:has([data-selected])]:rounded-r-md [&:has([data-selected][data-outside-month])]:bg-gray-100/50 [&:has([data-selected][data-selection-end])]:rounded-r-md',
+		className
+	)}
+	{...$$restProps}
+>
+	<slot />
+</RangeCalendarPrimitive.Cell>