+++ packages/core-components/src/lib/atoms/shadcn/select/select-label.svelte
@@ -1,0 +1,13 @@
+<script>
+	import { Select as SelectPrimitive } from 'bits-ui';
+	import { cn } from '$lib/utils';
+
+	/** @type {string | undefined | null} */
+	let className = undefined;
+
+	export { className as class };
+</script>
+
+<SelectPrimitive.Label class={cn('px-2 py-1.5 text-sm font-semibold', className)} {...$$restProps}>
+	<slot />
+</SelectPrimitive.Label>