+++ packages/core-components/src/lib/molecules/sql-console/SqlConsole.svelte
@@ -3,10 +3,10 @@
 </script>
 
 <script>
-	/** @typedef {import("@evidence-dev/query-store").QueryStore} QueryStore */
-	/** @typedef {import("./sqlConsole.action.js").SqlConsoleArgs} SqlConsoleArgs */
+	/** @typedef {import('@evidence-dev/query-store').QueryStore} QueryStore */
+	/** @typedef {import('./sqlConsole.action.js').SqlConsoleArgs} SqlConsoleArgs */
 	import { sqlConsole, buildAutoCompletes } from './sqlConsole.action.js';
-
+	import { slide } from 'svelte/transition';
 	import { onMount } from 'svelte';
 
 	import { Button } from '../../atoms/button';
@@ -18,7 +18,7 @@
 	/** @type {boolean} */
 	export let hideErrors = false;
 	/** @type {string} */
-	export let initialQuery = '';
+	export let initialQuery = "select 'ABC' as category, 123 as num, 26400000 as sales_usd";
 	/** @type {boolean} */
 	export let showResults = true;
 	/** @type {boolean} */
@@ -32,7 +32,7 @@
 
 	/*
 	 - Try to fix autocomplete
-	 	- non-namespaced?
+		 - non-namespaced?
 		- keyword updates
 	 */
 
@@ -76,11 +76,12 @@
 	);
 </script>
 
+<h1 class="markdown">SQL Console</h1>
 <section
-	class="px-4 py-2 bg-white flex flex-col gap-2 min-h-[8rem]"
+	class="px-0 py-2 bg-white flex flex-col gap-2 min-h-[8rem]"
 	on:click={() => editor?.focus()}
 	on:keydown={(e) => e.key === 'Enter' && editor?.focus()}
-	aria-roledescription="Code Editor"
+	role="none"
 >
 	<div
 		bind:this={editor}
@@ -109,8 +110,9 @@
 					on:click={() => {
 						currentQuery = editorQuery;
 						showResults = true;
-					}}>Submit</Button
-				>
+					}}
+					>Submit
+				</Button>
 			</div>
 		{/if}
 	</div>
@@ -121,7 +123,7 @@
 
 	<!-- Result View -->
 	{#if showResults}
-		<div>
+		<div transition:slide|local>
 			<DataTable data={$data} />
 		</div>
 	{/if}