+++ packages/core-components/src/lib/organisms/layout/header/KebabMenu.svelte
@@ -1,0 +1,69 @@
+<script>
+	import { Button } from '../../../atoms/shadcn/button';
+	import * as DropdownMenu from '../../../atoms/shadcn/dropdown-menu';
+	import { Icon } from '@steeze-ui/svelte-icon';
+	import { Settings, _3dCubeSphere, Link, Dots, Table, Prompt } from '@steeze-ui/tabler-icons';
+	import { showQueries } from '@evidence-dev/component-utilities/stores';
+	import { dev } from '$app/environment';
+
+	const beforeprint = new Event('export-beforeprint');
+	const afterprint = new Event('export-afterprint');
+	function print() {
+		window.dispatchEvent(beforeprint);
+		setTimeout(() => window.print(), 0);
+		setTimeout(() => window.dispatchEvent(afterprint), 0);
+	}
+</script>
+
+<DropdownMenu.Root>
+	<DropdownMenu.Trigger asChild let:builder>
+		<Button builders={[builder]} variant="ghost" size="sm" class="px-1" id="layout-kebab">
+			<Icon src={Dots} class="h-6 w-6" />
+		</Button>
+	</DropdownMenu.Trigger>
+	<DropdownMenu.Content class=" w-44 text-xs">
+		<DropdownMenu.Group>
+			<DropdownMenu.Item on:click={print}>
+				Print PDF
+				<DropdownMenu.Shortcut>⌘P</DropdownMenu.Shortcut>
+			</DropdownMenu.Item>
+			<DropdownMenu.Item
+				on:click={() => {
+					showQueries.update((val) => !val);
+				}}
+			>
+				{$showQueries ? 'Hide ' : 'Show '} Queries
+			</DropdownMenu.Item>
+		</DropdownMenu.Group>
+		{#if dev}
+			<DropdownMenu.Separator />
+			<DropdownMenu.Group>
+				<DropdownMenu.Item href="/settings" el="a">
+					Settings
+					<DropdownMenu.Shortcut><Icon src={Settings} class="w-4 h-4" /></DropdownMenu.Shortcut>
+				</DropdownMenu.Item>
+				<DropdownMenu.Item href="/settings/#deploy" el="a">
+					Deploy
+					<DropdownMenu.Shortcut><Icon src={_3dCubeSphere} class="h-4 w-4" /></DropdownMenu.Shortcut
+					>
+				</DropdownMenu.Item>
+				<DropdownMenu.Item href="/explore/schema" el="a">
+					Schema Viewer
+					<DropdownMenu.Shortcut>
+						<Icon src={Table} class="h-4 w-4" />
+					</DropdownMenu.Shortcut>
+				</DropdownMenu.Item>
+				<DropdownMenu.Item href="/explore/console" el="a">
+					SQL Console
+					<DropdownMenu.Shortcut>
+						<Icon src={Prompt} class="h-4 w-4" />
+					</DropdownMenu.Shortcut>
+				</DropdownMenu.Item>
+				<DropdownMenu.Item href="https://docs.evidence.dev" target="_blank" rel="noreferrer" el="a">
+					Documentation
+					<DropdownMenu.Shortcut><Icon src={Link} class="h-4 w-4" /></DropdownMenu.Shortcut>
+				</DropdownMenu.Item>
+			</DropdownMenu.Group>
+		{/if}
+	</DropdownMenu.Content>
+</DropdownMenu.Root>