+++ packages/core-components/src/lib/unsorted/viz/core/ECharts.svelte
@@ -53,6 +53,7 @@
 />
 
 <div
+	role="none"
 	class="chart-container"
 	on:mouseenter={() => (hovering = true)}
 	on:mouseleave={() => (hovering = false)}
@@ -119,10 +120,11 @@
 				stroke-width="2"
 				stroke-linecap="round"
 				stroke-linejoin="round"
-				><rect x="3" y="3" width="18" height="18" rx="2" /><circle cx="8.5" cy="8.5" r="1.5" /><path
-					d="M20.4 14.5L16 10 4 20"
-				/></svg
 			>
+				<rect x="3" y="3" width="18" height="18" rx="2" />
+				<circle cx="8.5" cy="8.5" r="1.5" />
+				<path d="M20.4 14.5L16 10 4 20" />
+			</svg>
 		</DownloadData>
 		{#if data}
 			<DownloadData
@@ -177,6 +179,7 @@
 			padding: 0;
 		}
 	}
+
 	.chart {
 		-moz-user-select: none;
 		-webkit-user-select: none;